backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
torch==2.9.1
torchvision==0.24.1
numpy==2.4.1
opencv-python==4.13.0.90
pillow==12.1.0
ultralytics==8.4.7
requests==2.32.5
python-multipart==0.0.21
pydantic==2.12.5
pydantic-core==2.41.5